+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Enabled parallel computation for Leave-One-Out values
+  [PR #406](https://github.com/appliedAI-Initiative/pyDVL/pull/406)
 - Fixes to parallel computation of generic semi-values: properly handle all
   samplers and stopping criteria, irrespective of parallel backend
   [PR #372](https://github.com/appliedAI-Initiative/pyDVL/pull/372)